// src/modules/candidates/types/candidate.types.ts
export interface ICandidateDetails {
    name: string;
    did: string;
    surname: string;
    dob: Date;
    birthplace: string;
    photo: string;
    email: string;
    phoneNo: string;
    emergencyContactName: string;
    emergencyContactRelationship: string;
    emergencyContactPhoneNo: string;
    ProjectSelectedFor: string;
    joiningDate: Date;
    jobType: string;
    formerEmployee: string;
    workingHours: string;
    workingMode: string; 
    previousExperience: string;
    professionOfCandidate: string;
    educationLevel: string;
    leanredGermanHow: string;
    recruitmentRepName: string;
    operaitonsRepName: string;
    englishLevel: string;
    germanLevel: string;
    comments: string;
    doc: any[];

}